\name{loadAnnotation}
\alias{loadAnnotation}
\title{Loads Annotation File}
\description{
Loads and checks an annotation file.
}
\usage{
loadAnnotation(fileName)
}
\arguments{
\item{fileName}{Name of file}
}
\details{
Annotation file need four columns: idNum, first, last, and overlap.
}
\value{Returns a data frame}
\references{See the book chapter O. Will (**) in **}
\author{Oliver Will \email{owill4@yahoo.com}}
\examples{
# No self contained example
}
\keyword{manip}