\name{loadInsertions}
\alias{loadInsertions}
\title{Load Genome Annotation File}
\description{
Loads a list of insertion locations. 
}
\usage{
loadInsertions(fileName)
}
\arguments{
\item{fileName}{Name of the file}
}
\details{Loads a list of insertion locations created in a transposon 
mutagenesis library.}
\value{Returns a data frame}
\references{See the book chapter O. Will (**) in **}
\author{Oliver Will \email{owill4@yahoo.com}}
\examples{
# No self contained example
}
\keyword{manip}