\name{sampleAnnotation}
\alias{sampleAnnotation}
\docType{data}
\title{Annotation for a Hypothetical Prokayote}
\description{
This dataset has the annotation for a hypothetical bacterium.
}
\usage{
data(sampleAnnotation)
}
\format{ 
A data frame containing 4 columns with 10 rows.
}
\source{Randomly generated.}
\references{See the book chapter O. Will (**) in **}
\author{Oliver Will \email{owill4@yahoo.com}}
\keyword{datasets}