\name{sampleInsertions}
\alias{sampleInsertions}
\docType{data}
\title{Insertions for a Hypothetical Clonal Library}
\description{
Insertion locations for a simple random mutagenesis library example.
}
\usage{
data(sampleInsertions)
}
\format{ 
A data frame containing 1 column with 20 rows.
}
\source{Randomly generated.}
\references{See the book chapter O. Will (**) in **}
\author{Oliver Will \email{owill4@yahoo.com}}
\keyword{datasets}